+++ vitest.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
   plugins: [react()],
   test: {
     environment: 'jsdom',
-    exclude: [...configDefaults.exclude, 'e2e/**'],
+    exclude: [...configDefaults.exclude, 'tests/e2e/**', 'e2e/**'],
     alias: {
       '@': path.resolve(__dirname, './')
     },
@@ -20,6 +20,7 @@ export default defineConfig({
         '**/*.test.ts',
         '**/*.test.tsx',
         'scripts/**',
+        'tests/e2e/**',
         'e2e/**',
       ],
     },